Another step to port the schedules view.
Don't use KMenu::setCheckable anymore since the doc says that it is not necessary and it's ignored.
Removed the KMenu::setItemChecked logic since now it's implemented in KMenu.
Name the list tab layout and use it to add the K3ListViewSearchLineWidget which I don't now why is missing the clear button.
Fixed the icon on the filter button.

svn path=/trunk/playground/office/kmymoney/; revision=1002241