BUG: 216499
Fix this by using the fontScaleFactor (only changes the size of the fonts) instead of the zoomFactor (which changes the size of all elements in the page images and fonts - and does not accept values < 100). I think this is the same with what we had on KDE3 and it works.

svn path=/trunk/playground/office/kmymoney/; revision=1056913