BUG: 217965
Second part of the fix for the investment transactions.
Add the missing labels. On the KDE3 version it worked in another way. The cell text for the non-edit widgets were drawn even when in edit mode but that didn't caused a problem like on Qt4 where for the labels only the text is drawn if the autoFillBackground property is false (default). So instead of setting autoFillBackground to true to draw the whole area of the label add the missing labels and don't draw the cell text when we're in edit mode.

svn path=/trunk/playground/office/kmymoney/; revision=1060761