This fixes the following wrong behaviour in the split transaction editor.
If the user left the category of the split empty and entered the split the currency conversion dialog would pop-up (setup to convert the account currency to the account currency) and after a useless currency conversion the operation would finish without any data entered.

The above behaviour was replaced by showing an information (with the don't show me again option) that the split needs to have a category in order to enter it and switches the edit focus to the category editor so that, after acknowledging the information, the user can select a category to finish the enter operation.

svn path=/trunk/playground/office/kmymoney/; revision=1079582