Fix some unit test linking issues on msvc.
Replace the '#define' of all access qualifiers to public with proper 'friend class' declarations for unit tests. Some compilers (like msvc) include the access qualifiers in the mangled method name so when we try to link the unit test we will fail because we are looking for a public method in a library where that method is actually private or protected.
Place the 'friend class' declarations in a non-intrusive way. This means that each testable object should have the KMM_MYMONEY_UNIT_TESTABLE macro in it's body. This macro expands to nothing if it's not previously defined. When writing a unit test define this macro before including any kmymoney header file to have the value 'friend class NameOfTheTestClass' (developer doc update to come).

svn path=/trunk/extragear/office/kmymoney/; revision=1146310