With the recent addition of choice of decimal symbol, a number of unit tests have been added to deal with the complexities of data formats, and to make easier the addition of further tests, should they be necessary.
Also, with the new tabbed UI, there are separate tabs for banking and investment, but just one, common table.  When switching tabs, the table gets cleared, but if the user accidentally clicks the wrong tab, it is necessary to reselect and reload the file.  I have added logic to issue a warning to guard against this accidental choice.
Further, I have added an '*' to the label of whichever is the active tab, again to remind the user.

svn path=/trunk/extragear/office/kmymoney/; revision=1224966