BUG: 265728 Sorry for the multiple changes (again) astyle keeps getting in my way: 1. Whitespace changes - astyle's fault 2. Removal of takeAxis-deleteAxis - if it's not needed by bar charts it shouldn't be needed by line charts (the chart will delete the axis) 3. The actual bugfix - this is a longer story see bellow
I managed to reproduce the bug described here by running the following steps (note that it will only work on a chart without a legend):
a. Press configure of a custom line chart without a legend (more items the the report setting)
b. Press cancel without changing anything
c. observe that the charts content is not visible
d. there are multiple ways to make it visible, resize the window, toggle chart/report/chart, etc.

The bottom line is that KDChart doesn't handle well charts without legends. So the fix involves changing a feature we already have.
Right now if the number of items was greater then a configuration setting we just hid the legend. Change this to always show the legend with a maximum of items given by the configuration setting... I hope this change feels right to everyone (otherwise we should hide the legend when the setting is 0 and fix KDChart to work without a legend)
This was a long description :)

svn path=/trunk/extragear/office/kmymoney/; revision=1231894