BUG: 270752
This bug was introduced by revision 1207610. Since that code sets the mouse button to Qt::RightButton the selectItem call from KGlobalLedgerView::selectEmptyTransaction would not go trough the normal codepath (marked by the '// we get here when called by application logic' comment) but trough the codepath handling the selection when the right button is pressed. But since that code was not working since the port to KDE4 (already fixed in Register::contextMenuEvent) the last item of the register would not get selected and thus provoking this bug.
Remove the invalid/unused right button handling in selectItem and reset the mouse button and the modifier on right click (Register::contextMenuEvent).

svn path=/trunk/extragear/office/kmymoney/; revision=1233105