BUG: 273770
Replace all usage of qFatal with qWarning in KMyMoney. I mean I've heard of assertive programming but this pushes it over the edge. qFatal('Can't open file') - I prefer letting the user know that there was an error and keep running to bringing down the whole application (it provides a feeling of instability) over a file open failure.

If I went too far with this change sorry but I prefer discussing re-adding the qFatal calls on a point-by-point basis. In most cases it's a shortcut for poor error handling.

svn path=/trunk/extragear/office/kmymoney/; revision=1233113