During QIF import of investments, categories of the form "Category:Sub-category" are created as a single category with no hierarchy.  In mymoneystatementreader.cpp, the routine d->interestId(t_in.m_strInterestCategory)) was found not to recognise a category:sub-category structure already existing, and would create a new category named like 'category:sub-category' without the hierarchy.  When the categoryToAccount() routine was substituted, this recognised and found the correct existing sub-account, but did not create one if none existed.  MyMoneyStatementReader::Private::nameToId was rewritten to handle category sub-accounts, recognising existing ones and otherwise creating them.
Handling of IntInc transactions did not allow for a security to be named, nor for the presence of quantities and prices.  Also, it did not allow for a fee to be described as IntInc. 

To assist with the handling of 'L' records which were indicating a category,
changes were made to use any existing brokerage account to supply/receive 
any monies.  If no brokerage account already existed, the record would be 
left flagged as missing an assignment.  This then highlighted a need for the ability to edit IntInc transactions, in order to specify an appropriate import account, and therefore included adding that action type to the UI.

svn path=/trunk/extragear/office/kmymoney/; revision=1237738