BUG: 280117
While testing an unreleased version of the CSV plugin, a crash occurred during import of an investment file. The crash occurred when selecting a column for the memo field, when that column had already been allocated.

The cause was an uninitialised variable.  

The same problem exists in the stable version of the plugin.

svn path=/trunk/extragear/office/kmymoney/; revision=1247434