BUG:276322 REVIEW:6791
If an existing cash dividend transaction is selected, the form shows widgets
for Date, Amount, Total and Status, which is what is expected.
If that transaction is then opened for editing, additional widgets show for
Security, Quantity and Price, which aren't expected for a cash dividend.
If, while the transaction is open for editing, 'Dividend' is reselected, the
additional fields disappear.  Similar behaviour occurs for Buy transactions.

Also, in numerous places, the likes of - 
QWidget* w = haveWidget(*it_s);
    if (w) 
      w->hide();
were found to produce unreliable results.

Also, When a Dividend transaction is opened for editing, the fee category field is
active, but there is no label.  If a fee category is entered, the fee amount
field becomes active - again no label, and if an amount is entered, the total
reflects this change.  However, when the transaction edit is accepted, the fee
amount has been lost.

Also, fees have been removed from Dividends, and interest income has been removed from 
Buy and Sell, as these seem most unlikely combinations.

I've noticed various other anomalies in the various investment types, to do
with unnecessary fields, missing labels and one where it's possible to enter a
value for a fee without a category, and which doesn't get retained.

svn path=/trunk/extragear/office/kmymoney/; revision=1254077