Do not allow the children widgets of the splitter to be collapsed.
This is needed in the payees and budgets view where the splitter has
two children and the data rendered on the right depends on the data
selected on the left so it does not make sense to hide (collapse)
any of the two widgets.

The setChildrenCollapsible call after restoreState is needed for users
that have an already saved state since the childrenCollapsible property
is saved with the state.

BUG: 284957