Scheduled transactions can't be skipped when entering them automatically at start-up or before reconciliation. Fix this by splitting slotScheduleSkip into slotScheduleSkip that works on m_selectedSchedule and skipSchedule that does the actual work and takes the schedule as a parameter (just like slotScheduleEnter/enterSchedule).
Use skipSchedule in enterSchedule if 'Skip' was selected instead of slotScheduleSkip.

As a side note the schedule is read again and again based on the id from the engine. Some of those calls could be
unnecessary so there is space for improvement here but I did not want to touch that so that I can safely backport
this fix to the stable branch.

BUG: 288647