Fixed the fact that select all was not updating the register until a paint was being triggered by the system (mouse over or something similar that caused the widget to be marked as dirty).
This problem is caused by the fact that the actual TableWidget which
is used to render the register is just a "painting shell". The actual
data is being kept in the register items which are used to paint the
register. The selection was correctly set in those items but no repaint
was triggered. Solved this by calling the view's selectAll method.

Also remove the unimplemented repaintItems method since it should not be needed
anymore.

The whole register would need a complete rewrite but we don't have the
resources yet to do that.

BUG: 283234