Fix some issues when Finance::Quote is used and perl is not available.
That causes QProcess to exit with the error signal instead of the finished signal and that causes kmymoney
to kepp processing events forever in the while loop of WebPriceQuote::quoteSourcesFinanceQuote which, in turn,
will cause it to hang after the application is closed.

Also properly initialize m_financeQuoteScriptPath before using it.