BUG: 294093 - Make sure that the header state is restored correctly.
Explanation (I figured this out by debugging Qt code):
Do this by setting the model of the view *before* the header state is being restored.
This is necessary because in some cases a layout of the items is triggered which causes
the header to ask for the column count from the model which is 0 if the model is not set
and this causes the restored section sizes to be lost after the model is finally set.