BUG: 286503 BUG: 294046
Attempt to reconnect to the database (once) on accidental disconnects. If this still throws, we need to notify the user and reset the UI to the same state as "kmymoney -n".

Add a test case to reproduce the problem in the storage layer.
(cherry picked from commit 4da92dfa03209670bddb7801926a25741c0ad6ec)