Use a safer way to set the minimum width of the account combo.
Using QFontMetrics::maxWidth() was a bad idea because some fonts could
contain really wide characters (with a very low probability of appearing
in the account names) - see the bug that is fixed.

Instead of the width of the widest character in the font use the width of
the 'W' character which will produce a much more predictable result.

Even the Qt implementation uses the width of a know character ('x') to
compute a kind of minimum size.

BUG: 295883