Don't show date popup on MAC OS
REVIEW: 18293

When the date is updated, a passive popup is shown displaying the full
date. This is not ideal on OSX because it steals focus from the date
input box, and when focus is restored the cursor is moved back to the
start of the input. This makes entering dates very annoying. Hence we
don't show the popup anymore on OSX to improve usability.
(cherry picked from commit a2a69ed9b60b0b426a8cee7a1f6e7ac7ad37fdc8)