Refactoring of matching a transaction-under-import
This commit does not introduce functional changes (to my and reviewers'
best knowledge).

The parts of code in TransactionMatcher class responsible for finding a
match for transaction under import have been extracted to dedicated
classes: ExistingTransactionMatchFinder (searches for a matching
transaction in the ledger) and ScheduledTransactionMatchFinder (searches
for a matching transaction in the schedules.

Also some bits of transaction import process (MyMoneyStatementReader)
have been refactored.

REVIEW: 107137