Improve the resizing of the register and the transaction form.
The transaction form was not correctly resizing the Value2 column
when the data of the KMyMoneyEdit exceeded the size of the widget.
The size of the push button (when present) was not considered when
the width was computed based on the cell text.

The same issues were fixed for the inline transaction editor.

The register resizing was improved in the following ways:
- the number fields is no longer limited
- the details column can no longer be shrinked to a size smaller then
it's needed to render the data it contains

BUG: 312816
REVIEW: 112989