Fixed a bug leading to crashes on removing onlineJobs from outbox
During the commit to remove the jobs signals are emitted. An invoked
slot caused a throw (onlineJobModel::data()) by requesting the
just removed onlineJob.

As the MyMoneyFileTransaction had m_needRollback == true a rollback was
done. But before emitting the signals MyMoneyFile d->m_inTransaction
was set to false causing another excepiton in
MyMoneyFile::rollbackTransaction().

Now the first excepiton is catched and the model was changed to prevent
the view to request in inconvenient situations (but even if this does
not work, first sentence applies).