Fix precision of price returned from price entry dialog
The returned price was truncated by the global price precision which
could cause all kinds of trouble when an amount was entered but the
price precision was not large enough to yield the same result in the
equation  result = amount / price * price;

Now the price is returned without precision loss in case it is based
on an amount (method is selected by the user in the dialog). In the
case the price is entered it is returned with the global price
precision in effect.

BUG: 205254
BUG: 325953
BUG: 300965