Performance optimization for MyMoneySeqAccessMgr::price().
Added testcases for this method.

According to valgrind this was the most expensive method of the
MyMoneySeqAccessMgr class which, in turn, was the most expensive
KMyMoney class considering the execution time.

After this price() moves bellow the transactionList() and
calculateBalance() methods and MyMoneySeqAccessMgr moves bellow
clases like MyMoneySplit and MyMoneyTransactionFilter.