Remove the code that messes with the file group identifier.
I think that code is not necessary anymore if it's stated purpose
is to preserve the groupid of the file. I've tested that the group id
is preserved anyway without that code and chwon would fail anyway
for a user which was not owner of the file but had write privileges
(in a multiuser environment).