Fix the alignment of the amounts in the ledger.
The alignment of the amounts (should be right) was broken since
the cells were rendered using a QTextDocument (since the tags
feature was added). I've tried to fix the QTextDocument alignment
using QTextDocument::setDefaultTextOption but it does not seem
to work as expected, until we find out why only the rich text
(containing tags) will be rendered using QTextDocument the rest
of the cell text is rendered as before.