Add Merge Payee Option. Patch by Felipe F. Tonello.
This patch implements the operation to merge payees. This is done by
reusing the payee delete slot, modifying it to support merge operation
and add support to it in KPayeeReassignDlg dialog.

The main difference between delete a payee and merging a payee is that
when merging the user can merge to a new payee. So the modification
was done to support this use-case.

REVIEW: 110212