Account ordering by name on the home page is now locale aware.
Since QMap does not allow setting the comparator function for each
instance of the map (qMapLessThanKey is for the whole class) we droped
using it just to obtain a collection ordered by the account name.
Use an account list instead and sort it using the proper comparator
before rendering it.

BUG: 335129