Preparing for frameworks.
Raise the cmake minimum required version to 2.8.6 (released in
10/2011). Did this so we could use the cmake automoc feature. Since
this feature is used we could remove all included moc files. See
this link [1] for a nice description of the feature.

Not all included moc files could be removed since we have some really
big dependency issues between our static libraries and using automoc
could mess up the functionality of the --as-needed linker flag.

Removed the vi markers since they we not present in all files anyway.

[1] https://blogs.kde.org/2011/11/01/cool-new-stuff-cmake-286-automoc