Preparing for Qt5.
Fix the tests that were failing on Qt5 because they were comparing
two XML fragments with different attribute orders. It seems that in
Qt5 attribute order is not stable anymore which is fine considering
the XML specifications but has the downside of messing up a possible
textual diff between two versions of the same kmy file.