Preparing for Qt5. Fix QDate usage in reports.
PivotTable::columnValue() was using an invalid QDate instead of a
"really old" reference date. This might have worked using Qt4 but
in Qt5 the QDate::daysTo() documentation states the following:

"Returns 0 if either date is invalid."

So use the start date of the report as a reference instead.
This fixes the last failing testcase. It's good to have unit tests
otherwise it would have been real hard to spot this.