Add feature to manually setup OFX app version
Since the application version changes from time to time (once a year?)
and we don't monitor this change, I added a feature that allows the user
to manually setup the application version for Quicken and Money. This
entry is marked as "(Expert)" version in the combo box. Once selected,
the edit field for the version becomes visible.