Let the build system generate the export header files.
This means that the definition of the KMM_MYMONEY_UNIT_TESTABLE
macro must be moved to another file since until now it was in the
export header of the mymoney library.

Note that the unknown '-std=c99' option had to be removed otherwise
the visibility check of the compiler would fail. See this thread:

http://mail.kde.org/pipermail/kde-frameworks-devel/2014-May/015954.html