Allow the user to edit details of investments with initial balance.
The opening balance transaction id should be cleared when the opening
balance is not editable. Otherwise we will try to remove an empty
transaction while updating the balance transaction which would cause
an error.

If the opening balance transaction of categories should not be edited
while editing the details hide the field also when creating a new
category.

BUG: 297223