Don't allow items of different types to be part of the same selection.
Check the typeids of the pointed to types instead of the pointer
types. This bad check was causing the correctly written code to fail.
With this fixed there is not way to change the reconlied state of a
scheduled transaction from the ledger thus the error will not appear
anymore.

BUG: 335122
(cherry picked from commit c89ebb9578222df5f84492a6a910d574b68a5839)