Changed payeeIdentifier system
payeeIdentifiers are not passed by shared pointers but with a new
payeeIdentifier class which does nearly the same as the shared pointer.
Only one difference: it removes the need to decide between copying the
pointer and the data. In the long run this will prevent a lot of
problems. This simplified the code in several places as well.

The old payeeIdentifier class is now payeeIdentifierData.

To enable this changes a lot of code was adapted and some bugs were
fixed.