Fix the forecast income and expenses reports.
The problem was that when the forecasted balance of a month was
requested the last day of the month was used instead of the first
day as MyMoneyForecast::calculateScheduledMonthlyBalances() sets
these balances on the first day.

This also implies that this kind of report can only have monthly
columns to be able to display valid data because of the structure
of the MyMoneyForecast object used. This is enforced by the reports
configuration UI.

BUG: 295239