Fix a nasty split editor bug.
Because of the way the endEdit method was written the endEdit
operation could be executed without actually ending the edit
(missing category). But code that was calling endEdit expected
that it allways would end the edit thus sometimes edit widgets
could be obtained on two rows.

Once in this state accepting both editors would cause a crash.

Also removed a workaround which would end the edit on a resize.
Instead don't do the operation that was resizing the rows while
in edit.

BUG: 336258