Fix the net worth report which was adding liabilities incorrectly.
I just noticed today that this was broken. The fix for this bug broke
it. The total should be computed differently for income and expense
vs. assets and liabilities forecast reports. I restored the old
condition and added a special condition to check only income and
expense forecast reports.

Alvaro if you have a better fix for this feel free to do it.

BUG: 295239