Made MyMoneyAccount interface to receive account number future safe
The interface is the same as MyMoneyPayeeIdentifierContainer. But it
does not inherit from it anymore. This is a step backwards but as long
as the actuall data is stored in MyMoneyAccount there is no good or
safe way to use MyMoneyPayeeIdentifierContainer.

This are only internal changes, the ui is not affected.