Remove the OFX investment transactions amount sign inversion hack.
As the author of the hack states he's not sure why this was necessary.
This hack also caused a lot of workarounds in MyMoneyStatementReader
so it definitely should be removed. I've tried to keep the current
behaviour in MyMoneyStatementReader but since I'm not familiar with
investment transaction these should be double checked.

BUG: 333522
REVIEW: 120153