Fix the payment method initialization in schedule editor.
Because the payment method filtering slot was connected before the
editor got setup it was being called during setup. This was wrong
since during setup the editor could emit operationTypeChanged with
various operations causing the filtering method to change a valid
initial value beacause it would not match the operation type.

Fixed this by connecting the slot after the editor was initialized.

BUG: 332267