Improve the previous ledger rendering related commits.
It seems that our font settings were useless because we were making
them before adding the legend. Chart::addLegend was overwriting our
font settings with sizes 20 and 24 (relative sizes which meant that
the fonts get larger if the rendered area is increased).

Now we set the font size after adding the legend so it is set
correctly to the size of the general font (an absolute value) so
the other size related workarounds are no longer necessary.

BUG: 265728
BUG: 290820