Database backend does not store row count in a table anymore
MyMoneyStorageSql stored the row count of several tables in the table
kmmFileInfo. But this is error-prone and bad style.

Now this information is read from the database directly. Usually a
database caches the row count of tables anyway. Also all tables have an
index so even a count(*) is really fast. The result is still cached
within MyMoneyStorageSQL as it did before.

REVIEW: 119647