Add tool to convert GNUCash templates to KMyMoney
I wrote this little tool a long time ago to convert template files
from GNUCash format to KMyMoney format as best as I could and used
it to convert the first set of templates.

Now people are asking for it, so here it is. I have not tested it
since 2008 so YMMV since the GNC format might have changed in the
meantime.

BUG: 339659