Update model completely when account data changes
Selecting to include a category in the tax report by marking the
checkbox in the account dialog presents a check mark in the category
view. Turning off the feature by unselecting the checkbox in the
account dialog did not remove the check mark in the category view.
This fixes the behavior.

BUG: 339693