Removed readFileInfo() and writeFileInfo() from upgradeToV7()
MyMoneyStorageSql::upgradeToV7() does not use readFileInfo() and
writeFileInfo() anymore.

This should prevent an issue if this metod is used to upgrade the
database to a version higher than 7 as readFileInfo() and
writeFileInfo() assume a database of the most recent version. This
will cause errors (=> crash) if a column in kmmFileInfo was
renamed or added between version 7 and the final version.

REVIEW: 120499