Fix the validation of investment transactions.
Don't ajust the values to the smallest account fraction of the
currency. This will cause roundings that will lead to the reported
validation error.

Because at 0.5 AlkValue::RoundRound can round up/down depending on
the result it could be that one split it rounded up while the other
is rounded down leading to a 0.1 error while summing them.

Having a value which is more precise than the smallest account
fraction should not cause any issues since the UI should render
the value according to this fraction anyway.

If rounding is necessary a fix rounding method should be applied
for all splits which are going to be summed up.

BUG: 303026
(cherry picked from commit 1506b0ccacb9d4efa2b6aaeb725c2e73f5507d83)