Remove the too harsh transation entering restrictions.
Now a transaction can be entered if at least one of the
following fields payee, category, amount is valid. This
was the old behavior and it seems better to be more
permissive with the user as long as no harm is done.

Attached the commit to the original bug report which seems
invalid since it requests these restrictions.

BUG: 314955