Fix the currency conversion dialog for really big/small conversion rates.
If the to value was equal to zero when the dialog was constructed a bad
(inverse) conversion rate was applied. This code was there since version
1.5 (on CVS) of this file. I don't understand why it was added that way
but now it's fixed.

Also fixe the prices dialog to display the prices using the configured
price precision.

Now in case that the value after conversion becomes zero the transaction
can't be entered.

BUG: 334564
(cherry picked from commit 7f61ffb728b58bbb905bba87e184e221a7d9858a)