Properly import OFX investment transactions.
Since fixing a previous issue the amount is no longer negative for
investment transactions so, to obtain the total, the fees need to be
added (have the same sign as the amount).

Now the provided file is imported without triggering the missing
assignment error. This was caused by miscalculated values, the shares
were properly imported.

BUG: 339192