Fix transaction entry when starting from transfer tab.
When the transfer tab was being disabled the next tab was being
set as the active tab automatically without considering currently
entered data. This would mess up the transaction as described in the
report.

To fix this, before disabling the transfers tab, if it is the current
tab, the proper deposit/withdrawl tab is selected.

BUG: 300697