Optimize the MyMoneyForecast object, now it uses less CPU cycles.
I really don't understand why the set of forecast accounts was
maintained as a map from the account id to the account id.

Changed this structure with a set of account ids representing
forecast accounts. Also perform some other operator[] related
optimizations.

After this the MyMoneyForecast object consumes one fourth of the
CPU cycles it used to consume while loading a KMyMoney file and
starting with the Home view.