Take advantage of the implicit sharing available in AlkValue.
Since the values 1 and -1 seem to be used pretty often (in loops)
it's worth to have static values for these constants to benefit from
implicit sharing in the places where these values are necessary.

The sharing of value 0 is already provided by AlkValue so we don't
need a static value for 0.

REVIEW: 120818