Refactor the contact fetching code into a single class.
This avoids code duplication and makes it easier to transform
the kdepimlibs dependency into an optional one. If kdepimlibs
is not detected the user will not be able to load contact data
from the address book (big deal...).

It also makes it easier to fetch contact data from other parts of the
code, like for a payee.

I'm not sure about putting this into the kmm_mymoney library but since
it only contains a data source and we don't have another lib for these
kind of objects it ended up here.

REVIEW: 119207