Fix the missing interest category for 'Buy' activities.
This fix is based on Allan's patch which is attached to the bug report
that I incorrectly thought to be fixed. I did some refactoring
because the readability of the investment activities was horrible. We
really should not have that kind of code in the application.

BUG: 276322
(cherry picked from commit 404d6279c74b8e93696f774fd04ac55d29307325)