Remove Yugoslavian Dinar
The Yugoslavian Dinar is not in use for almost 12 years, so there
is no need to create it in new files. The replacement in Serbia is
the Serbian Dinar which is already created for new KMyMoney files.
If a user uses a file created with a KMyMoney version < 4.5.96 the
Serbian Dinar (RSD) needs to be created manually.

BUG: 342275
(cherry picked from commit b3b15ce04d904f63d64b5103ebe725845ddac2dc)