Changed label for serach bar in ledger from "Search" to "Filter"
The field is actually more filter than search.

This also fixes the bug "In Ledger: Keyboard shortcut for Search (Alt-E)
does not work".

I could not change the label in scheduled transactions view because it
is hard coded in the library. But this should be done as well.

BUG: 338296
FIXED-IN: 4.7.2
REVIEW: 121803
(cherry picked from commit bf29b86bcb28c9a1837f351a0cb83f791c779eda)