Ensure no KBanking job is left in intenal queue
If a user wants to update an account but selects "Cancel" in the
dialogue where the first date for the transactions to download is
selected, the internal KBanking queue could be filled with other jobs
but not executed. As the outbox tab for KBanking will be removed that is
not an option anymore.

This fixed an intransparent behaviour: If you selected "Cancel" in that
dialogue before, the execution was not aborted is all cases.