Database version is set after creating tables
So the upgrade routine is not executed starting from version 0 - which
is useless as the database should be the most recent version anyway.

Also removed the version param from createTables(int version). It was
not used and was totaly useless as createTables() calls upgradeDb().

This is also a workaround for bug 341304 as the upgradeToV1() won't be
called after creation of tables anymore. The bug persists if the user
has a version 0 database.

CCBUG: 341304
REVIEW: 121288