Fix any initial state of the loan editor represented by group buttons.
Use QAbstractButton::click instead of QAbstractButton::animateClick
to set the initial state of some values. This is needed because when
animateClick is used the value is only set after 100 ms thus possibly
overwriting any initial values set by KEditLoanWizard::loadWidgets
which is called immediately after creating the wizard pages (thus
before the click would actually occur).

This fixes the state of loan type ('lending' or 'borrowing') in the
wizard.

BUG: 343348