Feature to copy splits from another transaction
This feature allows to create splits based on an existing transaction.
Simply select multiple transactions where one has a set of splits (can
be multiple) and the others have non (unassigned). The start this action
from the transaction context menu in the ledger view and the unassigned
transaction will have the same set with the same amounts as the
original.

In case the source transaction is a simple transaction (referencing only
two accounts) the value and share amounts are adjusted automatically to
the value present in the unassigned transaction. In case the two
accounts referenced by the transaction are based on different
currencies, the price of the source transaction is used to calculate the
share value of the newly created split.