Added traces for runtime of database manager tests
Since the tests of the database manager object are timing out on the KDE
CI infrastructure I added some output of the times that are taken by the
individual teststeps and the total runtime after each. This will be
removed once we have figured out how to reduce the runtime
significantly.