Added context menu to outbox
Added the obligatory context menu with "new", "edit" and "remove"
options to the outbox for onlineJobs.

This breaks with a KMyMoney design concept where the actions connect to
a slot in KMyMoneyApp. Naturally the edit and remove slots should be in
the onlineJobs view (as that is the place where the edit and remove
takes place). So the action is connected to the view directly.

REVIEW: 122545