Speed up database manager unittests
I was able to speed up the database manager tests by keeping an empty
database with all the tables already in place in a second file and
simply copying it over before each test step. Locally, I could reduce
the time needed for the tests to a third (before: ~160s, after: ~55s).