Add a stage to the consistency check to fix transaction post dates.
If the transaction is moved to another post date the price information
is moved along with it for investments transactions. This fixes the
reported issue, saving the file once will fix all price issues.

Also add the possibility to save the consistency check log to a file
or copy it into the clipboard.

Make sure that the user enters a valid transaction date by checking
that the transaction date is after the opening date of each and every
account involved in the transaction. Until now we only checked that
the transaction date is after the opening date of the account in which
is being entered.

Remove the check for "Sell" action since there is no such action.

BUG: 313793
REVIEW: 122746
(cherry picked from commit dbd547a503e14e146d7f9d2b1b0781f05ad3e0b6)