Properly handle loan payments from accounts using a different currency
If a payment account that has a different currency than the currency
of the loan is selected in the loan wizard, the resulting scheduled
transaction would not have the proper value in the currency of the
payment account.

Fixed this by requesting the user to perform the currency conversion
from the amount of the payment in the currency of the loan to the
amount in the currency of the payment account. This conversion is done
using the standard mechanism provided by the currency calculator.

After this was fixed I observed that the amount of the scheduled
transaction was not being rendered properly in the homepage. Fixed
this by using the same amount formating that is used in the ledger.

BUG: 325793
REVIEW: 123177
(cherry picked from commit d87427c4e8fad8226c1f1a80bbb094d9cc59a2ab)

Conflicts:
	kmymoney/views/khomeview.cpp