Fix the loan wizard losing the currency of the loan.
If a loan is created in a different currency than the base currency
and a new payee is created for the loan then the currency of the
loan ends up being the base currency instead of the originally
selected currency.

This bug was caused by some code which didn't made any sense (set an
invalid selection in the currency combo upon file change) so I just
removed it to fix the problem.

BUG: 345669
(cherry picked from commit aa330f9412c90f5161ff5b91f330cbc031a3f8c8)